Introduce public API function opk_open_metadata()
That function has to be called before any call to opk_read_param().
It opens the next meta-data file present on the OPK, and returns
its filename. If no other meta-data file can be found, it returns
NULL. The subsequent calls to opk_read_param() will return the
meta-data specified on the opened file.