unsqfs: Rewrote sparse file unpacking
The previous code was overly complex, probably because it originated
from a tool that supported multithreaded decompression and output to
sparse files.

It also contained bugs, such as block->buffer->data being freed in a
situation where block->buffer is NULL. Also the delayed writing of
zero-bytes was done while the buffer pointer was already pointing to
the first block after the gap.