unsqfs: Removed additional newlines from TRACE and ERROR macros
The extra newline would help if the terminal contains a partial line,
but chances of that are small. Meanwhile, an error is likely to trigger
multiple error messages from unsqfs since at different levels relevant
context info from that level is printed. Having these lines that belong
to the same error separated by newlines is confusing.