unsqfs: Allocate decompressed metadata per block
Previously, a single memory area was allocated for all metadata.
Allocating it per block is more efficient if we don't end up
decompressing all of it (a future improvement).
The problem though is that inode data can cross block boundaries, so
care must be taken to memcpy from two separate blocks then; this is
done in the read_metadata() function.