Changed opk_sqfs_extract_file() to return buffer length as well
The same change should be applied to opk_extract_file() too, but this
first step can be done without changing the library's interface. The
reason why buffer length must be output is that it is unsafe to just
assume that whatever data is passed in the buffer is complete:
a truncated file should not crash the application.