Avoid poking into extracted metadata buffer
This eliminates the need for the hack in opk_sqfs_extract_file() that
allocated one byte more such that the terminating zero char could be
inserted there. The hackish part of this being that unsqfs.c contained
knowledge of libopk.c, which violates the layered design.

This also opens the way for opk_sqfs_extract_file() returning a
read-only buffer in the future, which could avoid copies in some cases
by returning direct pointers to internal buffers.

In this commit copies were added in the form of strndup() calls.
However, I'm planning to offer next_param() as a public function in the
upcoming API update, which means the caller would decide whether to use
strndup() or not. Note that the strings are only valid until the next
metadata file is processed, so callers that want to keep the strings
already need to make copies.