Switch to using libini to parse the .desktop meta-data
Using libini simplifies the code of libopk by much. While it
adds a dependency to libopk, libini is extremely portable and
shouldn't cause any problem for portability.

Furthermore, libini provides the ability of reading meta-data
outside the main [Desktop Entry] section; as a result, additional
application actions could be implemented for OPKs, as defined
by the Freedesktop .desktop entry specification, to increase the
field of possibilities.