Return integers instead of booleans in functions of the public API
Previously, it was necesary to read the values of the pointers
passed as reference to opk_read_pair to know whether or not a
parse error occured, or to detect that no more pairs can be found,
as the function would return 'false' on both cases.

Now, we return an integer that gets the value -1 on error,
1 when a pair has been read, and 0 otherwise.

As we want to keep compatibility with applications compiled as C89,
although we use C11 for the library implementation, and as "bool"
is different in C and C++, we decided to get rid of boolean
values in the public API; as such, opk_extract_file now returns
an integer instead of a boolean, which equals -1 on error and 0
otherwise.