Change prototype of API function opk_open_metadata()
It now returns an integer instead of a pointer to the meta-data
file's name; the pointer passed by reference as the second
argument of the function will be used instead to return the
filename.

This allows us to return proper error codes, and to properly detect
errors (as NULL was returned on error but also if no more meta-data
file was to be found).