Added Python binding of function opk_extract_file
The Python function 'extract_file(opk, file)' can now be used.
Its first argument should be a string corresponding to the
filename of the OPK; the second argument should be a string
corresponding to the name of the file to extract. That function
returns an object of type 'StringIO' which can then be used
like any other file in Python, with read/seek/tell/close ...